appointment_site/requirements.txt
flask
flask_sqlalchemy
gunicorn
psycopg2-binary
pytz
python-dotenv
sqlite3
datetime